Tugas Kelompok/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.9.1
python-dotenv==1.0.1
PyJWT==2.9.0
